refactor: move solver greedy utilities out of __init__.py
Move _get_machine_intervals_for_task, _perform_task_interval_allocation,
_allocate_task, _create_schedule, and _schedule_by_order from the
solver package __init__ into a dedicated solver.greedy module. Update
all five callers (base_solver, dummy_solver, stochastic_solver,
genetic_solver, cp_sat_solver) to import from the new path. No
behavior change; preparation for Phase 2's instance-analysis split.
