+++ frost_planner/solver/stochastic_solver.py
@@ -12,8 +12,8 @@ from frost_planner.core.objective import (
     calculate_objective_value,
 )
 from frost_planner.core.schedule import ScheduledTask
-from frost_planner.solver import _create_schedule, _schedule_by_order
 from frost_planner.solver.base_solver import BaseSolver
+from frost_planner.solver.greedy import _create_schedule, _schedule_by_order
 
 
 class StochasticSolver(BaseSolver):
